+++ packages/flame/CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## [next]
+ - Add `SpriteAnimationWidgetBuilder`
+
 ## [1.0.0-releasecandidate.12]
  - Fix link to code in example stories
  - Fix RotateEffect with negative deltas