+++ packages/flame/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [next]
  - Add `SpriteAnimationWidgetBuilder`
+ - Add `SpriteWidgetBuilder`
 
 ## [1.0.0-releasecandidate.12]
  - Fix link to code in example stories