+++ packages/flame/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [next]
  - Add `SpriteAnimationWidgetBuilder`
  - Add `SpriteWidgetBuilder`
+ - Add `SpriteButtonBuilder`
 
 ## [1.0.0-releasecandidate.12]
  - Fix link to code in example stories