+++ packages/flame/CHANGELOG.md
@@ -4,6 +4,7 @@
  - Add `SpriteAnimationWidgetBuilder`
  - Add `SpriteWidgetBuilder`
  - Add `SpriteButtonBuilder`
+ - Add `NineTileBoxBuilder`
 
 ## [1.0.0-releasecandidate.12]
  - Fix link to code in example stories